shuttlexpress: add support for ShuttlePRO
The Contour ShuttlePRO uses the same protocol as the ShuttleXpress,
but has 13 buttons instead of 5.

This commit teaches the shuttlexpress driver to detect both the
ShuttleXpress and the ShuttlePRO.

The clicky wheel and the springy wheel and all buttons work on the
ShuttlePRO, but I haven not tested on the ShuttleXpress.

Still need to update docs, and maybe rename the component from
"shuttlexpress" to just "shuttle" or "contour" or something.  Also need
to update the manpage.

Oh, and since there can now be more than 9 buttons, maybe give the
buttons' hal pins two digits for their number, so they sort right.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>